Add docker for local db for testing

+++ backend/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Text, ARRAY
+from sqlalchemy import Column, Integer, String, Float, DateTime, Text, ARRAY, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.dialects.postgresql import ARRAY as PG_ARRAY
 import datetime
@@ -8,7 +8,7 @@ Base = declarative_base()
 class Issue(Base):
     __tablename__ = "issues"
 
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     number = Column(Integer, unique=True, index=True, nullable=False)
     title = Column(String, index=True, nullable=False)
     user = Column(String, nullable=False)
